add mybatis-generator-core-1.3.2

+++ ssm/src/main/java/com/ks/ssm/service/IUserService.java
@@ -3,6 +3,6 @@ package com.ks.ssm.service;
 import com.ks.ssm.domain.User;
 
 public interface IUserService {
-	public User getUserById(int userId);
+	public User getUserById(long userId);
 
 }
